Fixed wrong type for maximal domain size in tuple set (again, cardinality must be unsigned not signed)
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@7590 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/int/extensional/tuple-set.cc
@@ -116,7 +116,7 @@ namespace Gecode {
     min = real_min; max = real_max;
 
     // Domainsize
-    domsize = max - min + 1;
+    domsize = static_cast<unsigned int>(max - min) + 1;
 
     // Allocate tuple indexing data-structures
     tuples = heap.alloc<Tuple*>(arity);
